requirements.txt
numpy>=1.21.0
opencv-python>=4.5.0
requests>=2.26.0
python-dotenv>=0.19.0